split cli into different file solves the problem of missing /usr/bin/env python on some distributions the setuptools wrapper script generated during setup.py install handles this correctly
