+++ docktree/docktree.py
@@ -1,6 +1,4 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# PYTHON_ARGCOMPLETE_OK
 
 """
 Analyse dependencies of docker images.
@@ -11,12 +9,6 @@ from __future__ import absolute_import
 from .ImageLayer import ImageLayer
 
 import docker
-import sys
-import argparse
-try:
-    import argcomplete
-except ImportError:
-    pass
 
 
 def analyse_layers():
@@ -95,43 +87,3 @@ def remove_untagged_layers(layers):
         layers.pop(tag)
 
     return layers
-
-
-def parse_args(argv=sys.argv[1:]):
-    """
-    :param argv: arguments which we get called with
-    :return: the parsed arguments as object (see argparse doc)
-    :rtype: object
-    """
-
-    parser = argparse.ArgumentParser(description=__doc__)
-
-    parser.add_argument(
-        '-i',
-        '--intermediate',
-        action='store_true',
-        dest='print_intermediate',
-        default=False,
-        help='print intermediate (untagged) layers'
-    )
-
-    if 'argcomplete' in globals().keys():
-        argcomplete.autocomplete(parser)
-
-    return parser.parse_args(argv)
-
-
-def main():
-    """
-    parse arguments and run the desired action
-    """
-    args = parse_args()
-
-    layers = analyse_layers()
-    if not args.print_intermediate:
-        layers = remove_untagged_layers(layers)
-    heads = get_heads(layers)
-    print_tree(heads)
-
-if __name__ == '__main__':
-    main()
